+++ packages/nimble-components/src/text-field/styles.ts
@@ -63,6 +63,7 @@ export const styles = css`
         --ni-private-hover-bottom-border-width: 2px;
         border: 0px solid rgba(${borderRgbPartialColor}, 0.3);
         border-bottom-width: var(--ni-private-bottom-border-width);
+        gap: calc(${standardPadding} / 2);
         padding-bottom: calc(
             var(--ni-private-hover-bottom-border-width) -
                 var(--ni-private-bottom-border-width)
@@ -106,12 +107,40 @@ export const styles = css`
         border-bottom-color: ${borderHoverColor};
     }
 
+    :host([appearance='frameless'].clear-inline-padding) .root {
+        padding-left: 0px;
+        padding-right: 0px;
+    }
+
+    .root::before {
+        ${/* Empty string causes alignment issue */ ''}
+        content: ' ';
+        color: transparent;
+        width: 0px;
+        user-select: none;
+    }
+
+    :host([appearance='frameless'].clear-inline-padding) .root::before {
+        display: none;
+    }
+
+    .root::after {
+        ${/* Empty string causes alignment issue */ ''}
+        content: ' ';
+        color: transparent;
+        width: 0px;
+        user-select: none;
+    }
+
+    :host([appearance='frameless'].clear-inline-padding) .root::after {
+        display: none;
+    }
+
     [part='start'] {
         display: contents;
     }
 
     slot[name='start']::slotted(*) {
-        margin-left: calc(${standardPadding} / 2);
         flex: none;
     }
 
@@ -120,8 +149,7 @@ export const styles = css`
         font: inherit;
         background: transparent;
         color: inherit;
-        padding-top: 0px;
-        padding-bottom: 0px;
+        padding: 0px;
         height: calc(
             ${controlHeight} - ${borderWidth} -
                 var(--ni-private-hover-bottom-border-width)
@@ -129,8 +157,6 @@ export const styles = css`
         width: 100%;
         margin-top: auto;
         margin-bottom: auto;
-        padding-left: calc(${standardPadding} / 2);
-        padding-right: calc(${standardPadding} / 2);
         border: none;
         text-overflow: ellipsis;
     }
@@ -140,6 +166,15 @@ export const styles = css`
     .control:disabled,
     .control:active {
         outline: none;
+    }
+
+    .control:disabled {
+        ${
+            /* There's an issue with the input element where the ellipsized
+               overflowed text is blank when scrolled into view, so just clip instead.
+               See https://webcompat.com/issues/104481 */ ''
+        }
+        text-overflow: clip;
     }
 
     .control::selection {
@@ -174,7 +209,6 @@ export const styles = css`
     :host(.invalid) .error-content svg {
         height: ${iconSize};
         width: ${iconSize};
-        padding-right: 8px;
         flex: none;
     }
 
@@ -215,13 +249,12 @@ export const styles = css`
     }
 
     slot[name='actions']::slotted(*) {
-        margin-right: 8px;
         ${controlHeight.cssCustomProperty}: 24px;
     }
 `.withBehaviors(
-        appearanceBehavior(
-            TextFieldAppearance.Underline,
-            css`
+                appearanceBehavior(
+                    TextFieldAppearance.Underline,
+                    css`
             .root {
                 --ni-private-bottom-border-width: 1px;
                 padding-top: ${borderWidth};
@@ -237,10 +270,10 @@ export const styles = css`
                 --ni-private-bottom-border-width: 1px;
             }
         `
-        ),
-        appearanceBehavior(
-            TextFieldAppearance.Block,
-            css`
+                ),
+                appearanceBehavior(
+                    TextFieldAppearance.Block,
+                    css`
             .root {
                 background-color: rgba(${borderRgbPartialColor}, 0.1);
                 --ni-private-bottom-border-width: 0px;
@@ -286,10 +319,10 @@ export const styles = css`
                 --ni-private-bottom-border-width: 1px;
             }
         `
-        ),
-        appearanceBehavior(
-            TextFieldAppearance.Outline,
-            css`
+                ),
+                appearanceBehavior(
+                    TextFieldAppearance.Outline,
+                    css`
             .root {
                 --ni-private-bottom-border-width: 1px;
                 border-width: ${borderWidth};
@@ -300,10 +333,10 @@ export const styles = css`
                 top: calc(${controlHeight} - ${borderWidth});
             }
         `
-        ),
-        appearanceBehavior(
-            TextFieldAppearance.Frameless,
-            css`
+                ),
+                appearanceBehavior(
+                    TextFieldAppearance.Frameless,
+                    css`
             .root {
                 --ni-private-bottom-border-width: 0px;
                 padding-top: ${borderWidth};
@@ -319,21 +352,21 @@ export const styles = css`
                 --ni-private-bottom-border-width: 0px;
             }
         `
-        ),
-        themeBehavior(
-            css`
+                ),
+                themeBehavior(
+                    css`
             ${'' /* Light theme */}
             .control::-ms-reveal {
                 filter: invert(0%);
             }
         `,
-            css`
+                    css`
             ${'' /* Dark theme */}
             .control::-ms-reveal {
                 filter: invert(100%);
             }
         `,
-            // Color theme
-            Theme.Dark
-        )
-    );
+                    // Color theme
+                    Theme.Dark
+                )
+            );